website: Move to gatsby-theme-ocular (#121)
