--- website/gatsby-browser.js
@@ -1,1 +1,0 @@
-Object.assign(module.exports, require('ocular-gatsby/gatsby-browser'));
