--- website/gatsby-ssr.js
@@ -1,1 +1,0 @@
-Object.assign(module.exports, require('ocular-gatsby/gatsby-ssr'));
