+++ website/package.json
@@ -8,9 +8,9 @@
   "scripts": {
     "start": "yarn clean-examples && yarn clean && yarn develop",
     "clean": "rm -rf ./.cache ./public",
-    "develop": "gatsby develop --port=8001 --no-uglify",
-    "debug": "node --inspect-brk --no-lazy node_modules/gatsby/dist/bin/gatsby develop",
-    "build": "yarn clean-examples && yarn clean && gatsby build --no-uglify",
+    "develop": "gatsby develop",
+    "debug": "node --inspect-brk --no-lazy node_modules/gatsby/dist/bin/gatsby develop --no-uglify",
+    "build": "yarn clean-examples && yarn clean && gatsby build --no-uglify --prefix-paths",
     "serve": "gatsby serve",
     "deploy": "NODE_DEBUG=gh-pages gh-pages -d public",
     "clean-examples": "rm -rf examples ; (find ../examples -name node_modules -exec rm -rf {} \\; || true)"
@@ -21,9 +21,8 @@
     "react-table": "^6.10.0"
   },
   "devDependencies": {
-    "gatsby": "^2.15.0",
-    "gh-pages": "^2.0.1",
-    "ocular-gatsby": "^1.0.3",
-    "sharp": "^0.22"
+    "gatsby": "^2.18.0",
+    "gatsby-theme-ocular": "^1.1.0-alpha.4",
+    "gh-pages": "^2.0.1"
   }
 }
